Include ref to relevant detnet docs in introduction
The security model is largely based upon the different usecases found in
RFC8578. However, it is also important to point out that the dataplane
framework and architecture is the foundation upon which we build these
use-cases. These documents are important and the user should be told
early on to familiarize oneself with these. The same goes for the
information model.