Addressing major concern from Housley, missing jitter
Jitter is can be detrimental to real-time applications and is a common
source of unlinearities in control-loops. Being clear that jitter is a
know requirement and that this is a valid attack-vector is
important.

Adding jitter to abstract and introduction as requested.

Note: the original introduction was /identical/ to the one in RFC8655
"Deterministic Networking Architecture". Including jitter makes sense, but
it is worth noting that already accepted RFCs, and those that /define/
DetNet at that, omits jitter in the same section.