Adressing major concern from Housley - hash-based MAC
They way MAC works, is to generate a checksum of a message, often using
universal hashing and then sign this using a shared, secret key. With
HMAC, this the hashing is done using a generic cryptographic hash algorithms
(e.g. the SHA-familiy) and a private, secret key.

Both are acceptable as long as the secret key used to generate the MAC
is not compromised. Updating the text to drop "hash-based" and adding a
note about protecting the key used for signing the hash.