Addressing minor concern: using L3/L4
DetNet is based on IP (L3) and streams may include headers from
protocols higher up in the OSI-stack. Either way, the gist of this is to
find some set of attributes that an external can use to uniquely ascribe
different packets to individual streams. This may be tuples from the IP
header itself (DetNet fields or other fields) or if it is able to find
attributes from higher protocols.

It should be noted, albeit not relevant for /DetNet/ that the underlying
layer (e.g. TSN) will have its own set of unique attributes tied to a
specific stream.