Update RFC-references to be xml2rfc v3.5.0 compliant
Recent updates to the xml2rfc tool caused parsing of detnet-security.xml
to fail on missing references (404 from tools.ietf.org), probably due to
change of URL used to retrieve the references.

xml2rfc 2.15.5 seems to append .xml when this is missing, something
3.5.0 does not do. Appending .xml to references solves this issue, and
is backwards-compatible with 2.15.5.

Moving to xml2rfc 3.5.0 does introduce a few other changes, most notably
the date-formatting and initials of authors.
i.e.
 "Dec. 3 2020" -> "3 December 2020"
 "E. Grossman, Ed." -> "E.A.G. Grossman, Ed."