AUTH48 version of Draft 16
I replaced the pre-existing draft 16 version of detnet-security.xml with the new version that came back from the RFC editors as "rfc9055_auth48.xml". This version was reformatted and converted to xml2rfc v3 so basically no lines matched at all so essentially the existing text is all deleted and the new text added. Probably I ended up reformatting it (i.e. whitespace changes) when I did my editing, so next I will check in a version of this file with that reformatting, so that it will be a reasonable diff between that and my new changes (to address all of the editorial comments).